Add to package trackerbot
