+++ src/main/java/trackerbot/Events.java
@@ -1,15 +1,16 @@
+package trackerbot;
+
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
 
-public class Deadlines extends Task {
-    protected LocalDateTime deadline;
+public class Events extends Task {
+    protected LocalDateTime startDate;
+    protected LocalDateTime endDate;
 
-    public Deadlines(String description, String deadline) throws TrackerBotException {
+    public Events(String description, String startDate, String endDate) throws TrackerBotException {
         super(description);
-
-        //to add date recognition - accepts YYYY-mm-dd HHMM / MMM dd yyyy HHmm
         try {
             DateTimeFormatter format = new DateTimeFormatterBuilder()
                     .optionalStart()
@@ -17,14 +18,19 @@ public class Deadlines extends Task {
                     .appendPattern("[MMM dd yyyy HHmm]")
                     .optionalEnd()
                     .toFormatter();
-            this.deadline = LocalDateTime.parse(deadline, format);
+            this.startDate = LocalDateTime.parse(startDate, format);
+            this.endDate = LocalDateTime.parse(endDate, format);
         } catch (DateTimeException e) {
             throw new TrackerBotException("Invalid Date-Time. Please Use YYYY-MM-DD HHMM or MMM DD YYYY HHmm");
         }
     }
 
     public String toString() {
-        String formattedDeadline = this.deadline.format(DateTimeFormatter.ofPattern("MMM dd yyyy HHmm"));
-        return String.format("[D][%s] %s (by: %s)", this.getStatusIcon(), this.description, formattedDeadline);
+        DateTimeFormatter format = DateTimeFormatter.ofPattern("MMM dd yyyy HHmm");
+        return String.format("[E][%s] %s (from: %s to: %s)",
+                this.getStatusIcon(),
+                this.description,
+                this.startDate.format(format),
+                this.endDate.format(format));
     }
 }
