+++ src/main/java/trackerbot/Task.java
@@ -1,3 +1,5 @@
+package trackerbot;
+
 public class Task {
     protected String description;
     protected boolean isDone;
