+++ src/main/java/trackerbot/TrackerBot.java
@@ -1,3 +1,5 @@
+package trackerbot;
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Scanner;
