+++ src/main/java/trackerbot/TrackerBotException.java
@@ -1,3 +1,5 @@
+package trackerbot;
+
 public class TrackerBotException extends Exception {
     public TrackerBotException(String message) {
         super(message);
